Bring horizontal input back in to the event center due to its dependencies
