--- Assets/Code/Input/HorizontalInput.cs
@@ -1,14 +1,0 @@
-﻿
-
-namespace PQ
-{
-    public struct HorizontalInput
-    {
-        public enum Type { None, Left, Right }
-        public readonly Type value;
-        public HorizontalInput(Type value)
-        {
-            this.value = value;
-        }
-    }
-}
