--- Assets/Code/Input/HorizontalInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e62451404dbe2b42bd00fc6bf551dbd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
