+++ Assets/Code/PlayerSettingsInfo.cs
@@ -1,6 +1,5 @@
 ﻿using UnityEngine;
 using System.Diagnostics.Contracts;
-using PQ.Common;
 
 
 namespace PQ
